### 2020-07-05 #2443 @whedon
[PRE REVIEW]: Grama: A Grammar of Model Analysis
**Submitting author:** @zdelrosario (<a href="http://orcid.org/0000-0003-4676-1692">Zachary del Rosario</a>)
**Repository:** <a href="https://github.com/zdelrosario/py_grama" target ="_blank">https://github.com/zdelrosario/py_grama</a>
**Version:** v0.1.3
