### 2020-07-05 #2443 @danielskatz
@openjournals/dev  - something seems broken here - the reference check does not return
