### 2020-07-05 #2443 @danielskatz
👋 @zdelrosario - while we get the process started, you might fix some of the bib issues that you can see in the generated PDF related to cases (e.g., "AIAA scitec 2020 forum") by protecting words or letters in {}s and any other issues you see (such as Ward et al. missing the book/proceedings).  Once you have made changes in the .bib or .md files, add `@whedon generate pdf` as a new comment here to see how things look.
