### 2020-07-05 145affa
fix del Rosario et al. (2020); fix JMLR; fix Ward et al. (2018)
