+++ paper.bib
@@ -2,8 +2,8 @@
 @Comment --------------------------------------------------
 @inproceedings{delRosario2020design,
   title={When are Design Allowables Conservative?},
-  author={del Rosario, Zachary and Fenrich, Richard W and Iaccarino, Gianluca},
-  booktitle={AIAA Scitech 2020 Forum},
+  author={{del Rosario}, Zachary and Fenrich, Richard W and Iaccarino, Gianluca},
+  booktitle={AIAA {SciTech} 2020 {Forum}},
   pages={0414},
   year={2020}
 }
@@ -78,11 +78,11 @@
 @article{pedregosa2011scikit,
   title={Scikit-learn: Machine learning in Python},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={the Journal of machine Learning research},
+  journal={The Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
   year={2011},
-  publisher={JMLR. org}
+  publisher={JMLR.org}
 }
 
 @Comment Tidyverse
@@ -127,11 +127,16 @@
 
 @Comment Matminer
 @Comment --------------------------------------------------
-@inproceedings{ward2018matminer,
+@article{ward2018matminer,
   title={Matminer: An open source toolkit for materials data mining},
-  author={Logan T. Ward and Alex Dunn and Alireza Faghaninia and Nils E. R. Zimmermann and Saurabh Bajaj and Qi Wang and J. E Paredes Montoya and Jiming Chen and Kyle Bystrom and Maxwell T. Dylla and Kyle Chard and Mark Asta and Kristin A. Persson and Gerald Jeffrey Snyder and Ian T. Foster and Anubhav Jain},
-  year={2018}
+  author={Ward, Logan and Dunn, Alexander and Faghaninia, Alireza and Zimmermann, Nils ER and Bajaj, Saurabh and Wang, Qi and Montoya, Joseph and Chen, Jiming and Bystrom, Kyle and Dylla, Maxwell and others},
+  journal={Computational Materials Science},
+  volume={152},
+  pages={60--69},
+  year={2018},
+  publisher={Elsevier}
 }
+
 
 @Comment Comparative software
 @Comment --------------------------------------------------
