### 2020-07-06 #2443 @arfon
@danielskatz - looks like the DOI checker is failing to parse the lines that look [like this one](https://github.com/zdelrosario/py_grama/blob/master/paper.bib#L1-L2).

@zdelrosario - could you try removing the lines of your BibTeX file that start with `@comment`?
