### 2020-07-06 2a8154f
rm comments; add doi's
