+++ paper.bib
@@ -1,11 +1,10 @@
-@Comment Motivation
-@Comment --------------------------------------------------
 @inproceedings{delRosario2020design,
   title={When are Design Allowables Conservative?},
   author={{del Rosario}, Zachary and Fenrich, Richard W and Iaccarino, Gianluca},
   booktitle={AIAA {SciTech} 2020 {Forum}},
   pages={0414},
-  year={2020}
+  year={2020},
+  doi = {https://doi.org/10.2514/6.2020-0414}
 }
 
 @article{freeman2014active,
@@ -16,11 +15,10 @@
   number={23},
   pages={8410--8415},
   year={2014},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi = {https://doi.org/10.1073/pnas.1319030111}
 }
 
-@Comment Scikit packages
-@Comment --------------------------------------------------
 @article{scipy2020,
        author = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant},
          Travis E. and {Haberland}, Matt and {Reddy}, Tyler and
@@ -52,7 +50,8 @@
   year={2011},
   volume={13},
   number={2},
-  pages={22-30}
+  pages={22-30},
+  doi = {https://doi.org/10.1109/MCSE.2011.37}
 }
 
 @article{matplotlib2007,
@@ -62,7 +61,8 @@
   year={2007},
   volume={9},
   number={3},
-  pages={90-95}
+  pages={90-95},
+  doi = {https://doi.org/10.1109/MCSE.2007.55}
 }
 
 @InProceedings{mckinney-proc-scipy-2010,
@@ -82,11 +82,9 @@
   volume={12},
   pages={2825--2830},
   year={2011},
-  publisher={JMLR.org}
+  publisher={JMLR.org},
 }
 
-@Comment Tidyverse
-@Comment --------------------------------------------------
 @article{wickham2019tidyverse,
   title = {Welcome to the {tidyverse}},
   author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang
@@ -110,11 +108,10 @@
   author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
   booktitle={ELPUB},
   pages={87--90},
-  year={2016}
+  year={2016},
+  doi = {10.3233/978-1-61499-649-1-87}
 }
 
-@Comment dfply
-@Comment --------------------------------------------------
 @misc{kiefer2019dfply,
   author = {Katovich, Kiefer},
   title = {dfply},
@@ -125,8 +122,6 @@
   commit = {6a858f066602735a90f8b6b85106bc39ceadc282}
 }
 
-@Comment Matminer
-@Comment --------------------------------------------------
 @article{ward2018matminer,
   title={Matminer: An open source toolkit for materials data mining},
   author={Ward, Logan and Dunn, Alexander and Faghaninia, Alireza and Zimmermann, Nils ER and Bajaj, Saurabh and Wang, Qi and Montoya, Joseph and Chen, Jiming and Bystrom, Kyle and Dylla, Maxwell and others},
@@ -134,12 +129,11 @@
   volume={152},
   pages={60--69},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.commatsci.2018.05.018}
 }
 
 
-@Comment Comparative software
-@Comment --------------------------------------------------
 @techreport{adams2017sandia,
   title={Sandia Capabilities for Uncertainty Quantification.},
   author={Adams, Brian M},
