### 2020-07-06 #2443 @danielskatz
Looks like the problem has been solved.

@zdelrosario - there are some DOIs that you should probably add to your bib file (though note that whedon sometimes gets these wrong.  Also, the invalid ones need to have the "https://doi.org/" removed
