+++ paper.bib
@@ -4,7 +4,7 @@
   booktitle={AIAA {SciTech} 2020 {Forum}},
   pages={0414},
   year={2020},
-  doi = {https://doi.org/10.2514/6.2020-0414}
+  doi = {10.2514/6.2020-0414}
 }
 
 @article{freeman2014active,
@@ -16,7 +16,7 @@
   pages={8410--8415},
   year={2014},
   publisher={National Acad Sciences},
-  doi = {https://doi.org/10.1073/pnas.1319030111}
+  doi = {10.1073/pnas.1319030111}
 }
 
 @article{scipy2020,
@@ -40,7 +40,7 @@
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 
 @article{numpy2011,
@@ -51,7 +51,7 @@
   volume={13},
   number={2},
   pages={22-30},
-  doi = {https://doi.org/10.1109/MCSE.2011.37}
+  doi = {10.1109/MCSE.2011.37}
 }
 
 @article{matplotlib2007,
@@ -62,7 +62,7 @@
   volume={9},
   number={3},
   pages={90-95},
-  doi = {https://doi.org/10.1109/MCSE.2007.55}
+  doi = {10.1109/MCSE.2007.55}
 }
 
 @InProceedings{mckinney-proc-scipy-2010,
@@ -130,7 +130,7 @@
   pages={60--69},
   year={2018},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.commatsci.2018.05.018}
+  doi = {10.1016/j.commatsci.2018.05.018}
 }
 
 
@@ -146,5 +146,6 @@
   author={Marelli, Stefano and Sudret, Bruno},
   booktitle={Vulnerability, uncertainty, and risk: quantification, mitigation, and management},
   pages={2554--2563},
-  year={2014}
+  year={2014},
+  doi = {10.1061/9780784413609.257}
 }
