### 2020-07-07 1ff5f92
add option to set ignore_errors argument
