### 2020-07-09 #2443 @BastinRobin
I am interested!.

On Thu, Jul 9, 2020 at 10:40 AM Matthew Sottile <notifications@github.com>
wrote:

> 👋 @rodrigokataishi <https://github.com/rodrigokataishi>, @BastinRobin
> <https://github.com/BastinRobin>, @jaeyk <https://github.com/jaeyk>,
> @lesaffre <https://github.com/lesaffre> - the author of this submission
> to JOSS has indicated that you may be interested in reviewing this
> submission. If you are interested and willing to provide a review within
> the next few weeks I would greatly appreciate it. Please respond in this
> issue if you are willing and I will add you as a reviewer. Thank you!
>
