### 2020-07-09 #2443 @rodrigokataishi
I´m interested too.
Please let me know how to proceed.
Best,
r

On Thu, Jul 9, 2020 at 2:41 AM Bastin Robin <notifications@github.com>
wrote:

> I am interested!.
>
> On Thu, Jul 9, 2020 at 10:40 AM Matthew Sottile <notifications@github.com>
> wrote:
>
> > 👋 @rodrigokataishi <https://github.com/rodrigokataishi>, @BastinRobin
> > <https://github.com/BastinRobin>, @jaeyk <https://github.com/jaeyk>,
> > @lesaffre <https://github.com/lesaffre> - the author of this submission
> > to JOSS has indicated that you may be interested in reviewing this
> > submission. If you are interested and willing to provide a review within
> > the next few weeks I would greatly appreciate it. Please respond in this
> > issue if you are willing and I will add you as a reviewer. Thank you!
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/2443#issuecomment-655904288
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/AA24IR65S7DXAKHVV2S5KNDR2VGK5ANCNFSM4ORASUXQ
> >
> > .
> >
>
>
> --
> Regards,
>
> *Bastin Robin.J*
> bastinrobins@gmail.com
>
