### 2020-07-10 #2462 @whedon
[REVIEW]: Grama: A Grammar of Model Analysis
**Submitting author:** @zdelrosario (<a href="http://orcid.org/0000-0003-4676-1692">Zachary del Rosario</a>)
**Repository:** <a href="https://github.com/zdelrosario/py_grama" target ="_blank">https://github.com/zdelrosario/py_grama</a>
**Version:** v0.1.3
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.12725066.v2" target="_blank">10.6084/m9.figshare.12725066.v2</a>
