### 2020-07-17 aac98d1
reimplement replace w/ re.sub; add + test str_replace_all
