### 2020-07-17 b668789
test replace missed pattern
