+++ tests/test_string_helpers.py
@@ -59,11 +59,16 @@ class TestStringHelpers(unittest.TestCase):
         self.assertTrue(gr.str_replace("foofoo", "foo", "barbar") == "barbarfoo")
         self.assertTrue(gr.str_replace("barbar", "foo", "bar") == "barbar")
 
-        R = gr.str_replace(["foo"], "foo", "bar")
+        R = gr.str_replace(["foo", "nope"], "foo", "bar")
         self.assertTrue(R[0] == "bar")
+        self.assertTrue(R[1] == "nope")
 
     def test_str_replace_all(self):
         self.assertTrue(gr.str_replace_all("foofoo", "foo", "bar") == "barbar")
+
+        R = gr.str_replace_all(["foofoo", "nope"], "foo", "bar")
+        self.assertTrue(R[0] == "barbar")
+        self.assertTrue(R[1] == "nope")
 
     def test_str_sub(self):
         s_base0 = "foofoo"
