### 2020-07-17 1ac3021
implement and test str_split
