### 2020-07-17 8c8c86e
reset index on arrange
