### 2020-07-24 #2462 @rodrigokataishi
@mjsottile Dear Mathiew. I can´t edit the checklist. It appears to be closed.
Any help will be welcome. I already accepted the invitation. But I can not edit the list to confirm the items.
Thanks
