### 2020-07-24 #2462 @mjsottile
@rodrigokataishi That should fix your problem accessing the checklist.  Let me know if it doesn’t.
