### 2020-07-27 #2462 @mjsottile
@zdelrosario : Please fix the citations in the last section of the paper.  It appears that you use commas to separate multiple entries in a single citation block, which is causing the reference names to appear in the text instead of the correct citations.  Instead you should use semicolons.  See: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Once you've fixed that, please archive a tagged release of your submission on an archive site such as Zenodo or figshare, and post the DOI for the archive to this issue thread.  Finally, ensure that the version number for the submission at the top of this issue matches the version of the archived code.  If it does not, let me know what the correct version number should be so I can set it.
