### 2020-07-27 #2462 @kthyng
Hi @zdelrosario! I'm taking over to finish publication of this paper. I'm not so familiar with figshare, but can you edit the metadata on your archive there so that the title and author list match your paper for this publication? It's nice to have it consistent.
