### 2020-07-28 418980d
add linear drag model
