+++ grama/models/__init__.py
@@ -1,6 +1,7 @@
 from .cantilever_beam import *
 from .ishigami import *
 from .test import *
+from .trajectory_linear_drag import *
 from .linear_normal import *
 from .plane_laminate import *
 from .plate_buckling import *
