+++ grama/data/__init__.py
@@ -3,17 +3,6 @@
 Built-in datasets.
 
 Datasets:
-    df_stang: Aluminum alloy data from Stang et al. (1946). Columns:
-        thick (inches): Nominal thickness
-
-        alloy: Alloy designation
-
-        E (Kips/inch^2): Young's modulus
-
-        mu (-): Poisson's ratio
-
-        ang (degrees): Angle of test to alloy roll direction
-
     df_diamonds: Diamond characteristics and prices. Columns:
         carat:
 
@@ -35,9 +24,28 @@ Datasets:
 
         z:
 
+    df_stang: Aluminum alloy data from Stang et al. (1946). Columns:
+        thick (inches): Nominal thickness
+
+        alloy: Alloy designation
+
+        E (Kips/inch^2): Young's modulus
+
+        mu (-): Poisson's ratio
+
+        ang (degrees): Angle of test to alloy roll direction
+
+    df_trajectory_full: Simulated trajectory data. Columns:
+
+        t: Time since projectile launch (seconds)
+
+        x: Projectile range (meters)
+
+        y: Projectile height (meters)
+
 References:
     Stang, Greenspan, and Newman, "Poisson's ratio of some structural alloys for
     large strains" (1946) U.S. Department of Commerce National Bureau of Standards
 """
 
-from .datasets import df_stang, df_diamonds
+from .datasets import *
