### 2020-07-28 ae846ba
add eval_nls example using trajectory
