### 2020-07-28 75b18c8
assign results in test
