+++ tests/test_models.py
@@ -26,11 +26,11 @@ class TestModels(unittest.TestCase):
         md_trajectory_linear = models.make_trajectory_linear()
 
         ## Models evaluate
-        md_cantilever_beam >> gr.ev_nominal(df_det="nom")
-        md_ishigami >> gr.ev_nominal(df_det="nom")
-        md_linear_normal >> gr.ev_nominal(df_det="nom")
-        md_plane_laminate >> gr.ev_nominal(df_det="nom")
-        md_plate_buckling >> gr.ev_nominal(df_det="nom")
-        md_poly >> gr.ev_nominal(df_det="nom")
-        md_test >> gr.ev_nominal(df_det="nom")
-        md_trajectory_linear >> gr.ev_nominal(df_det="nom")
+        df_cantilever = md_cantilever_beam >> gr.ev_nominal(df_det="nom")
+        df_ishigami = md_ishigami >> gr.ev_nominal(df_det="nom")
+        df_ln = md_linear_normal >> gr.ev_nominal(df_det="nom")
+        df_plane = md_plane_laminate >> gr.ev_nominal(df_det="nom")
+        df_plate = md_plate_buckling >> gr.ev_nominal(df_det="nom")
+        df_poly = md_poly >> gr.ev_nominal(df_det="nom")
+        df_test = md_test >> gr.ev_nominal(df_det="nom")
+        df_traj = md_trajectory_linear >> gr.ev_nominal(df_det="nom")
