### 2020-07-28 f3d2089
add prototype fit_nls()
