+++ grama/__init__.py
@@ -93,3 +93,5 @@ from .eval_opt import *
 from .plot_auto import *
 from .tran_shapley import *
 from .tran_summaries import *
+
+from .fit_synonyms import *
