### 2020-07-28 #2462 @zdelrosario
Thanks @kthyng, @mjsottile , @BastinRobin , and @rodrigokataishi! This is my first time publishing with JOSS, and I'm super-impressed by your speed and process.
