### 2020-07-28 42db0fc
handle base model w/o name
