### 2020-07-28 4b67f40
add windowed trajectory data
